+++ lib/src/create_report.dart
@@ -595,6 +595,13 @@ Future<List<OutdatedVersionDescription>> computeOutdatedVersions(
             parsedVersion <= Version.parse(package.upgradable.version)) {
           continue;
         }
+        // It's not outdated, just mutually incompatible if allowed by the
+        // constraint, but still reported as outdated by `dart pub outdated`.
+        if (hostedDependency.version == null ||
+            hostedDependency.version.allows(parsedVersion)) {
+          continue;
+        }
+
         final publishingDateString = tryGetFromJson<String>(
                 version, 'published') ??
             // If the pub host doesn't provide a `published` time, we pretend it
@@ -785,7 +792,9 @@ Future<ReportSection> trustworthyDependency(PackageContext context) async {
               status = ReportStatus.failed;
             } else if (worst.status == OutdatedStatus.outdatedByPreview ||
                 worst.status == OutdatedStatus.outdatedByRecent) {
-              status = ReportStatus.partial;
+              // TODO(sigurdm): Could we find some way of indicating that
+              // points will be lost soon?
+              status = ReportStatus.passed;
             }
           }
         }
@@ -887,9 +896,7 @@ _Issue _explanationToIssue(Explanation explanation) =>
     _Issue(explanation.finding, suggestion: explanation.explanation);
 
 Future<ReportSection> _nullSafety(String packageDir, Pubspec pubspec) async {
-  // TODO(sigurdm): Currently we don't give any points for null-safety.
-  // Raise this after null-safety beta.
-  const maxPoints = 0;
+  const maxPoints = 20;
 
   _Subsection subsection;
   if (File(p.join(packageDir, '.dart_tool', 'package_config.json'))
@@ -902,18 +909,14 @@ Future<ReportSection> _nullSafety(String packageDir, Pubspec pubspec) async {
     if (pubspec.sdkConstraintStatus.hasOptedIntoNullSafety) {
       if (nullSafetyTags.contains('is:null-safe')) {
         subsection = _Subsection(
-            'Package and dependencies are fully migrated to null-safety, '
-            'and will be awarded additional points in a planned future revision '
-            'of the pub.dev points model.',
+            'Package and dependencies are fully migrated to null safety!',
             explanations.map(_explanationToIssue).toList(),
             maxPoints,
             maxPoints,
             ReportStatus.passed);
       } else {
         subsection = _Subsection(
-            'Package declares support for null-safety, but there are issues.\n\n'
-            'Packages with full null-safety support will be awarded additional '
-            'points in a planned future revision of the pub.dev points model.',
+            'Null safety support has one or more issues.',
             [
               ...explanations.map(_explanationToIssue).toList(),
               // TODO(sigurdm): This is no longer enough, because `pub outdated`
@@ -929,9 +932,7 @@ Future<ReportSection> _nullSafety(String packageDir, Pubspec pubspec) async {
       }
     } else {
       subsection = _Subsection(
-          'Package does not opt in to null-safety.\n\n'
-          'Packages with full null-safety support will be awarded additional '
-          'points in a planned future revision of the pub.dev points model.',
+          'Package does not opt in to null safety.',
           [
             _Issue(
               'Package language version (indicated by the sdk constraint '
@@ -948,7 +949,7 @@ Future<ReportSection> _nullSafety(String packageDir, Pubspec pubspec) async {
     subsection = _Subsection(
       'Unable to detect null safety',
       [
-        _Issue('Package resolution failed. Could not determine null-safety.',
+        _Issue('Package resolution failed. Could not determine null safety.',
             suggestion: 'Run `pub get` for more information.')
       ],
       0,
@@ -957,7 +958,7 @@ Future<ReportSection> _nullSafety(String packageDir, Pubspec pubspec) async {
     );
   }
   return _makeSection(
-    title: 'Support sound null-safety',
+    title: 'Support sound null safety',
     maxPoints: maxPoints,
     id: ReportSectionId.nullSafety,
     subsections: [subsection],